Support destination writer and color disabling
This PR resolves #3. It adds a `New` function that will create
a new handler while accepting a destination writer. It also
accepts an option to enable colorizing output.

I went with the [functional options pattern][fop] as it is very
common and works well.

The advertised default usage has not changed. That is, it still
writes to `stdout` with colorized output by default when utilizing
the `NewHandler` function to create the handler instance.